Move Value into its own file
This is also pretty undiscoverable and a pain to figure out.  Here's how to move code into its own file:

1. Create a file with the appropriate name in your module.
2. Update `mod.rs` to declare it as a module (optionally, publicly).
3. Update any uses to its new namespace.
4. Regret the exceptional ergonomics misstep that this feels like.

I suppose that Rustaceans feel like this is pretty low key and normal, but this wasn't readily discoverable and the rust-lang.org book didn't seem to cover it well enough that I understood it on first reading nor answer my questions when re-referenced.  All I really want to do is have files with single responsibility, but this is adding a lot of namespacing overhead and inconvenience without obvious merit.